int -> s64

+++ test-6/test-5-check.cpp
@@ -3,12 +3,12 @@
 #include <iostream>
 #include <vector>
 
-int expectedResult(int r, int s) {
+s64 expectedResult(s64 r, s64 s) {
     R5_ASSERT(r >= 1);
     R5_ASSERT(s >= 1);
 
-    int r_ = r;
-    int s_ = s;
+    s64 r_ = r;
+    s64 s_ = s;
     r = std::min(r_, s_);
     s = std::max(r_, s_);
 
@@ -19,8 +19,8 @@ int expectedResult(int r, int s) {
     }
 
     if (r == 3) {
-        const std::vector<int> results{-1, -1, -1, 6, 9, 14, 18, 23, 28, 36};
-        if (s >= int(results.size()) - 1) {
+        const std::vector<s64> results{-1, -1, -1, 6, 9, 14, 18, 23, 28, 36};
+        if (s >= s64(results.size()) - 1) {
             std::cerr << "I don't know R(" << r << "," << s << ")" << std::endl;
             return -1;
         } else {
@@ -29,8 +29,8 @@ int expectedResult(int r, int s) {
     }
 
     if (r == 4) {
-        const std::vector<int> results{-1, -1, -1, -1, 18, 25};
-        if (s >= int(results.size()) - 1) {
+        const std::vector<s64> results{-1, -1, -1, -1, 18, 25};
+        if (s >= s64(results.size()) - 1) {
             std::cerr << "I don't know R(" << r << "," << s << ")" << std::endl;
             return -1;
         } else {
